removed spark conf from swan_config
