+++ swan-upstream-chart/files/swan_config.py
@@ -10,7 +10,7 @@ Class handling KubeSpawner.modify_pod_hook(spawner,pod) call
 """
 
 
-class PodHookHandler:
+class SwanPodHookHandler:
     def __init__(self, spawner, pod):
         """
         :type spawner: swanspawner.SwanKubeSpawner
@@ -34,13 +34,6 @@ class PodHookHandler:
 
         # get eos token
         eos_secret_name = self._init_eos_secret()
-
-        # get hadoop token
-        hadoop_secret_name = None
-        if self._spark_enabled():
-            # cern customisation for spark clusters
-            hadoop_secret_name = self._init_hadoop_secret()
-            self._init_spark(pod_labels)
 
         if self._gpu_enabled():
             # currently no cern customisation required
@@ -77,10 +70,10 @@ class PodHookHandler:
             )
 
         # init pod affinity
-        self._init_pod_affinity(pod_labels)
+        self.pod.spec.affinity = self._init_pod_affinity(pod_labels)
 
         # init user containers (notebook and side-container)
-        self._init_user_containers(eos_secret_name, hadoop_secret_name)
+        self._init_user_containers(eos_secret_name)
 
         return self.pod
 
@@ -125,93 +118,7 @@ class PodHookHandler:
 
         return eos_secret_name
 
-    def _init_hadoop_secret(self):
-
-        cluster = self.spawner.user_options[self.spawner.spark_cluster_field]
-
-        if cluster == 'none':
-            return None
-
-        username = self.spawner.user.name
-        hadoop_secret_name ='hadoop-tokens-%s' % username
-
-        hadoop_token_base64 = ''
-        webhdfs_token_base64 = ''
-        k8suser_config_base64 = ''
-
-        if cluster == 'k8s':
-            hdfs_cluster = 'analytix'
-            try:
-                # Setup the user and generate user kube config
-                k8suser_config_base64 = subprocess.check_output(
-                    ['sudo', '/srv/jupyterhub/private/sparkk8s_token.sh', username], timeout=60
-                ).decode('ascii')
-            except Exception as e:
-                # if no access, all good for now
-                raise ValueError("Could not setup user on k8s")
-            try:
-                # Retrieve HDFS, YARN token for user
-                hadoop_token_base64 = subprocess.check_output(
-                    ['sudo', '/srv/jupyterhub/private/hadoop_token.sh', hdfs_cluster, username], timeout=60
-                ).decode('ascii')
-            except Exception as e:
-                # if no access, all good for now
-                raise ValueError("Could not get spark tokens")
-            try:
-                # Retrieve hdfs token for user
-                webhdfs_token_base64 = subprocess.check_output(
-                    ['sudo', '/srv/jupyterhub/private/webhdfs_token.sh', hdfs_cluster, username], timeout=60
-                ).decode('ascii')
-            except Exception as e:
-                # if no access, all good for now
-                raise ValueError("Could not get webhdfs tokens")
-        else:
-            try:
-                # Retrieve HDFS, YARN token for user
-                hadoop_token_base64 = subprocess.check_output(
-                    ['sudo', '/srv/jupyterhub/private/hadoop_token.sh', cluster, username], timeout=60
-                ).decode('ascii')
-            except Exception as e:
-                # if no access, all good for now
-                raise ValueError("Could not get spark tokens")
-            try:
-                # Retrieve hdfs token for user
-                webhdfs_token_base64 = subprocess.check_output(
-                    ['sudo', '/srv/jupyterhub/private/webhdfs_token.sh', cluster, username], timeout=60
-                ).decode('ascii')
-            except Exception as e:
-                # if no access, all good for now
-                raise ValueError("Could not get webhdfs tokens")
-
-        # Create V1Secret with eos token
-        try:
-            secret_data = client.V1Secret()
-
-            secret_meta = client.V1ObjectMeta()
-            secret_meta.name = hadoop_secret_name
-            secret_meta.namespace = swan_container_namespace
-            secret_data.metadata = secret_meta
-            secret_data.data = {}
-            secret_data.data['k8s-user.config'] = k8suser_config_base64
-            secret_data.data['hadoop.toks'] = hadoop_token_base64
-            secret_data.data['webhdfs.toks'] = webhdfs_token_base64
-
-            try:
-                self.spawner.api.read_namespaced_secret(hadoop_secret_name, swan_container_namespace)
-                exists = True
-            except ApiException:
-                exists = False
-
-            if exists:
-                self.spawner.api.replace_namespaced_secret(hadoop_secret_name, swan_container_namespace, secret_data)
-            else:
-                self.spawner.api.create_namespaced_secret(swan_container_namespace, secret_data)
-        except ApiException as e:
-            raise Exception("Could not create required hadoop secret: %s\n" % e)
-
-        return hadoop_secret_name
-
-    def _init_user_containers(self, eos_secret_name, hadoop_secret_name):
+    def _init_user_containers(self, eos_secret_name):
         """
         Define cern related secrets for spark and eos
         """
@@ -284,53 +191,6 @@ class PodHookHandler:
             )
         )
 
-        if hadoop_secret_name:
-            # pod volume to mount generated hadoop tokens and
-            # side-container volume mount with generated tokens
-            self.pod.spec.volumes.append(
-                # V1Secret for tokens without adjusted permissions
-                client.V1Volume(
-                    name=hadoop_secret_name,
-                    secret=client.V1SecretVolumeSource(
-                        secret_name=hadoop_secret_name,
-                    )
-                )
-            )
-            side_container_volume_mounts.append(
-                client.V1VolumeMount(
-                    name=hadoop_secret_name,
-                    mount_path='/srv/side-container/hadoop'
-                )
-            )
-
-            # define hadoop auth environment for the notebook container
-            notebook_container.env = self._add_or_replace_by_name(
-                notebook_container.env,
-                client.V1EnvVar(
-                    name='HADOOP_TOKEN_FILE_LOCATION',
-                    value='/srv/notebook/tokens/hadoop.toks'
-                ),
-            )
-            notebook_container.env = self._add_or_replace_by_name(
-                notebook_container.env,
-                client.V1EnvVar(
-                    name='KUBECONFIG',
-                    value='/srv/notebook/tokens/k8s-user.config'
-                ),
-            )
-            notebook_container.env = self._add_or_replace_by_name(
-                notebook_container.env,
-                client.V1EnvVar(
-                    name='WEBHDFS_TOKEN',
-                    value_from=client.V1EnvVarSource(
-                        secret_key_ref=client.V1SecretKeySelector(
-                            key='webhdfs.toks',
-                            name=hadoop_secret_name
-                        )
-                    )
-                ),
-            )
-
         # append as first (it will be first to spawn) side container which currently:
         #  - refreshes the kerberos token and adjust permissions for the user
         self.pod.spec.volumes.append(
@@ -378,33 +238,6 @@ class PodHookHandler:
         # assigning pod spec containers
         self.pod.spec.containers = pod_spec_containers
 
-    def _spark_enabled(self):
-        """
-        Helper function to determine if spark related configuration is necessary
-        raise exception if user has not access to the selected spark cluster
-        return True if spark cluster is selected and user has access to the selected spark cluster
-        return False if spark cluster is not selected
-        """
-
-        user_roles = self.spawner.user_roles
-        cluster = self.spawner.user_options[self.spawner.spark_cluster_field]
-
-        if cluster == "analytix" and "analytix" not in user_roles:
-           raise ValueError(
-              """
-              Access to the Analytix cluster is not granted. 
-              Please <a href="https://hadoop.web.cern.ch/" target="_blank">request access</a>
-              """)
-        elif cluster == "hadoop-nxcals" and "hadoop-nxcals" not in user_roles:
-           raise ValueError(
-              """
-              Access to the NXCALS cluster is not granted. 
-              Please <a href="https://wikis.cern.ch/display/NXCALS/Data+Access+User+Guide#DataAccessUserGuide-nxcals_access" target="_blank">request access</a>
-              """)
-        elif cluster != "none":
-            return True
-        return False
-
     def _gpu_enabled(self):
         """
         Helper function to determine if gpu is allowed for given spawn
@@ -421,149 +254,6 @@ class PodHookHandler:
         elif "cu" in lcg_rel:
             return True
         return False
-
-    def _init_spark(self, pod_labels):
-        """
-        Set cern related configuration for spark cluster and open ports
-        """
-        notebook_container = self._get_pod_container('notebook')
-        username = self.spawner.user.name
-
-        cluster = self.spawner.user_options[self.spawner.spark_cluster_field]
-        max_mem = self.spawner.user_options[self.spawner.user_memory]
-
-        if cluster == 'none':
-            return
-
-        spark_ports_service = "spark-ports" + "-" + username
-
-        # add basic spark envs
-
-        notebook_container.env = self._add_or_replace_by_name(
-            notebook_container.env,
-            client.V1EnvVar(
-                name='SPARK_CLUSTER_NAME',
-                value=cluster
-            )
-        )
-        notebook_container.env = self._add_or_replace_by_name(
-            notebook_container.env,
-            client.V1EnvVar(
-                name='SPARK_USER',
-                value=username
-            )
-        )
-        notebook_container.env = self._add_or_replace_by_name(
-            notebook_container.env,
-            client.V1EnvVar(
-                name='MAX_MEMORY',
-                value=max_mem
-            )
-        )
-
-        # add spark auth required env
-
-        if cluster == 'hadoop-nxcals':
-            auth_required = 'true'
-        else:
-            auth_required = 'false'
-
-        notebook_container.env = self._add_or_replace_by_name(
-            notebook_container.env,
-            client.V1EnvVar(
-                name='SPARK_AUTH_REQUIRED',
-                value=auth_required
-            )
-        )
-
-        # add spark config env
-
-        if cluster == 'k8s':
-            spark_conf_script = '/cvmfs/sft.cern.ch/lcg/etc/hadoop-confext/k8s-swan-setconf.sh'
-        else:
-            spark_conf_script = '/cvmfs/sft.cern.ch/lcg/etc/hadoop-confext/hadoop-swan-setconf.sh'
-
-        notebook_container.env = self._add_or_replace_by_name(
-            notebook_container.env,
-            client.V1EnvVar(
-                name='SPARK_CONFIG_SCRIPT',
-                value=spark_conf_script
-            )
-        )
-
-        # configure spark ports
-        try:
-            spark_ports_env = []
-
-            # Define some 6 random NodePorts on the cluster for spark using V1Service
-            service_template_ports = []
-            spark_ports_per_pod = 6
-            for port_id in range(1, spark_ports_per_pod + 1):
-                service_template_ports.append(
-                    client.V1ServicePort(
-                        name="spark-port-" + str(port_id),
-                        port=port_id
-                    )
-                )
-            service_template = client.V1Service(
-                api_version="v1",
-                kind="Service",
-                metadata=client.V1ObjectMeta(
-                    name=spark_ports_service
-                ),
-                spec=client.V1ServiceSpec(
-                    selector=pod_labels,  # attach this service to the pod with label {spark_pod_label}
-                    ports=service_template_ports,
-                    type="NodePort"
-                )
-            )
-
-            # Create V1Service which allocates random ports for spark in k8s cluster
-            try:
-                # use existing if possible
-                self.spawner.api.delete_namespaced_service(spark_ports_service, swan_container_namespace)
-                service = self.spawner.api.read_namespaced_service(spark_ports_service, swan_container_namespace)
-            except ApiException:
-                # not existing, create
-                service = self.spawner.api.create_namespaced_service(swan_container_namespace, service_template)
-
-            # Replace the service with allocated nodeports to map nodeport:targetport
-            # and set these ports for the notebook container
-            for port_id in range(len(service.spec.ports)):
-                name = service.spec.ports[port_id].name
-                node_port = service.spec.ports[port_id].node_port
-                service.spec.ports[port_id] = client.V1ServicePort(
-                    name=name,
-                    node_port=node_port,
-                    port=node_port,
-                    target_port=node_port
-                )
-
-                # Construct ports env for spark
-                spark_ports_env.append(str(node_port))
-
-                # Open proper ports in the notebook container to map nodeport:targetport
-                notebook_container.ports = self._add_or_replace_by_name(
-                    notebook_container.ports,
-                    client.V1ContainerPort(
-                        name=name,
-                        container_port=node_port,
-                        # this is needed - hadoop-yarn webapp crashes on ApplicationProxy UI
-                        host_port=node_port,
-                    )
-                )
-            self.spawner.api.replace_namespaced_service(spark_ports_service, swan_container_namespace, service)
-
-            # Add ports env for spark
-            notebook_container.env = self._add_or_replace_by_name(
-                notebook_container.env,
-                client.V1EnvVar(
-                    name='SPARK_PORTS',
-                    value=','.join(spark_ports_env)
-                )
-            )
-        except ApiException as e:
-            raise Exception("Could not create required user ports: %s\n" % e)
 
     def _init_pod_affinity(self, pod_labels):
         """
@@ -586,7 +276,7 @@ class PodHookHandler:
             )]
         )
         aff.pod_affinity = pod_affinity
-        self.pod.spec.affinity = aff
+        return aff
 
     def _get_pod_container(self, container_name):
         """
@@ -622,7 +312,7 @@ def modify_pod_hook(spawner, pod):
     :returns: dynamically customized pod specification for user session
     :rtype: client.V1Pod
     """
-    pod_hook_handler = PodHookHandler(spawner, pod)
+    pod_hook_handler = SwanPodHookHandler(spawner, pod)
     return pod_hook_handler.get_swan_user_pod()
 
 """
@@ -648,10 +338,6 @@ c.KeyCloakAuthenticator.oidc_issuer = 'https://auth.cern.ch/auth/realms/cern'
 c.KeyCloakAuthenticator.accepted_roles = set()
 c.KeyCloakAuthenticator.auto_login = True
 c.KeyCloakAuthenticator.admin_role = 'swan-admins'
-
-
-# https://jupyterhub-kubespawner.readthedocs.io/en/latest/spawner.html
-c.SwanKubeSpawner.modify_pod_hook = modify_pod_hook
 
 # Get configuration parameters from environment variables
 swan_container_namespace = os.environ.get('POD_NAMESPACE', 'default')
