+++ swan-upstream-chart/templates/config.yaml
@@ -6,6 +6,7 @@ metadata:
 data:
 {{ (.Files.Glob "files/options_form_config.json").AsConfig | indent 2 }}
 {{ (.Files.Glob "files/swan_config.py").AsConfig | indent 2 }}
+{{ (.Files.Glob "files/swan_spark_config.py").AsConfig | indent 2 }}
 {{ (.Files.Glob "files/private/cull_check_ticket.sh").AsConfig | indent 2 }}
 {{ (.Files.Glob "files/private/cull_delete_ticket.sh").AsConfig | indent 2 }}
 {{ (.Files.Glob "files/private/side_container_tokens_perm.sh").AsConfig | indent 2 }}
